requirements.txt
ipykernel==6.29.5
jupyter==1.1.1
linearmodels==6.1
matplotlib==3.9.2
nbformat==5.10.4
numpy==2.1.1
pandas==2.2.3
pyyaml==6.0.2
pytest==8.3.3
scipy==1.14.1
seaborn==0.13.2
statsmodels==0.14.4
